Preserve the FIM programming files in each variant folder (#82)
The pr-build-template folder is huge, but we really only need/want the blue_bits subfolder (~45MB) which contains the FIM's programming files. I'm leaving the pr-build-template copy in for now, but will eventually remove it and we'll be left with just the blue_bits folder in each variant. This allows different variants to use different FIMs.
Ran build-bsp.sh multiple times in the same place - the blue_bits folder get replaced each time the command is executed.