Fix memory-clock clock-group hierarchy assignments (#91)
### Description
The current n6001 master branch has problems with timing due to a hierarchy change in the memory-subsystems leading to an invalid clock-group assignment. This leads to recovery violations that can't be fixed by reducing the kernel-clock frequencies, and the kernel-clock being incorrectly reduced to a useless frequency.

### Tests run:
Manually modified the opencl_bsp.sdc file in-place using the Timing Analyzer GUI for a qri build that had timing violations, and all of the violations disappeared (they were all on CDC paths).