[Update] D5005 - leverage FIM's pr scripts for user-clock and gen-gbs #97
###  Description
Recently the N6001 ASP was updated to take advantage of some scripts included in the pr-build-template folder from the FIM that handled the user/kernel-clock manipulation and generation of the gbs file. This PR does the same for the D5005 variants.
This helps get us closer to merging all platforms and variants into common code.

## Tests run:
Successfully compiled a few designs (boardtest, simple-host-streaming, zero-copy-data-transfer) and tested in hardware.