[Update] Update path to syn_top in entry.tcl (#110)
### Description
Remove the hard-coding of the syn_top path in entry.tcl to prevent a change in the FIM hierarchy breaking the ASP's build/compilation flow. This change will detect the relative path to the syn_top folder and write it into a variable in a tcl file from setup-bsp.py, which is then sourced in entry.tcl and used to find the Quartus build location.

### Collateral (docs, reports, design examples, case IDs):
None.

###  Document Update Required? (Specify FIM/AFU/Scripts)
## Tests added:
## Tests run:
Compiled N6001 and D5005 sample designs successfully.