[Update] Commonize setup-bsp.py to find default qpf and qsf files (#120)
### Description
This change adds some intelligence to the setup-bsp.py script to find the appropriate qpf and qsf files from the pr-build-template folder and symlink them where the OneAPI build-flow expects them to be. Previously these files were hardcoded depending on whether the platform was d5005 or n6001, but that wasn't helpful when porting FIMs. This should simplify porting the ASP to different platforms (as long as the filenames are somewhat consistent).


### Collateral (docs, reports, design examples, case IDs):



- [ ] Document Update Required? (Specify FIM/AFU/Scripts)

### Tests added:


### Tests run:
Ran build-bsp.py for both n6001 and d5005, and then successfully compiled many oneapi-samples designs.