*[Fix]- Update header files to match changes to FIM #144
## Description
A recent change to the FIM's defines caused the local-memory defines in the ASP to be mishandled. This will hopefully go away with the move to qprs/ip files, but this is a fix for the rest of the Agilex variants for now. I accidentally only fixed one of the vh files with my earlier commit, while the rest of the vh files were only partially fixed (I moved the bug around instead of resolving it for those incomplete changes). I noticed this while using ASE with iseries-dk and the memory-interfaces in board.sv were weird.

### Tests run:
Compilation of a few designs for each variant. Opening the database in the Quartus GUI shows me that the ddr_board and ddr_channel IPs consume resources, whereas without the change those IPs were either missing or tiny. Also, the build-time increased drastically (it doesn't take much time to compile a kernel with no access to local memory).